backend/requirements.txt
# CodeCollab Backend - Python 3.13 compatible
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
supabase>=2.10.0
httpx>=0.28.0
